add package helm-ls-git
helm-ls-git is an awesome package that lets me quickly operate on any
git controlled files in the current repo. for things like staging it
is faster then magit because I don't need to load a diff. This is
especially true for files that are in different directories but are
similar. In magit they are spaced out because they are not in the same
part of the hierarchy, so I have to stage one, wait, stage the next
one, wait, etc. With helm-ls-git I can just narrow the files and stage
or commit them all at once.

Also in my limited testing $helm-browse-project is faster at listing
files then projectile. I will have to see if that holds true for large repos.