Move to a private fork of helm
I am doing this for two reasons.
1. I realized that I have a lot of code whoes job is to hack the helm
internals to make it behave like I want. This is to get around issues
that the helm maintainer refuses to add support for. It is much easier
to make these changes inside helm then to try and make them
externally.
2. I want to learn helm better so that I can use it more powerfully.
This will force me to spend more time in the code.