Fix bugs in find similar file function
if only one file was found, it would not display the path from the
point of divergence. This was fixed by adding the original file to the
common parent algorithm.
If the first file found was a symlink to the original file, it would
accept that and stop searching. This was fixed by chasing symlinks.