Fix error in RTL bman and add command history to generic command
we were still running vclp which is a DRC tool, however this relied on
hip_listgen, which we were excluding. This led to many weird errors.
Also sglp_legacy is now deprecated.

By adding history to the bman generic command, it makes it much easier
to run previous commands. Eventually I want to make this persistent.