Fix overlay bug in paren matching
This has been a long one. I randomly had this code lead to some
failures. I thought that it was related to ov not being defined so I
added this check, but it turns out it was an issue with lexical
scoping not being enabled. I was running into another issue where the
overlays were not being removed and I tracked it down to this
statement.