Remove local-env-mode
Boy did this give me some issues! This package (which I wrote) advises getenv to
fetch the local environment before calling the function. If the buffer was
remove we would use tramp to get the local env. However tramp would try to call
getenv, which would trigger local-env, which would trigger tramp again. This
lead to an infinite loop. Everytime I used tramp I hit this loop and basically got
a stack overflow.

I had to slowly work out which function was causing the issue by edebugging each
on and then wait for the overflow.