Fix undo tree not being enabled by default
This was a really hard bug to track down. There was a recent change in evil mode
where undo was not being forced on .They were calling turn-on-undo-tree-mode
instead of undo-tree-mode. The former checks to make sure that you are not
overriding the undo tree bindings. C-/ is a undo tree binding that we were
overriding in helm-company. This was causing undo-tree to not be enabled.