Remove literate config and simplify things
Using a literate config did not justify the extra overhead. You had to deal with
the time to tangle the init and file issues when it didn't work. This meant that
you were often loading org early, leading to slow startup. Also working with a
long chunk of code is just more convenient than having to deal with multiple org
blocks. An elisp buffer is also faster than an large org buffer. We can still
use outline mode to jump to headings just like we did with org.

I thought it would be nice to share a specific parts of my literate config with
people on github, but I have maybe only done that twice in six years. Not worth
it.